+++ lib/core/constants/strings.dart
@@ -1,0 +1,7 @@
+class Strings {
+  // App
+  static const appTitle = 'Instant';
+  static const appDesc = 'Instagram clone';
+  static const appVersion = 'Version: 1.0.4';
+  static const appDev = 'Developed by Chetan R Billava';
+}